+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Add `userId` to the recommendation input.
+
 ## [1.6.4] - 2022-01-28
 
 ### Fixed